requirements.txt
numpy
psycopg2-binary
pandas
Flask
uwsgi